migrate our composition functions to two separate scripts for clarity
